Added timer class

+++ src/Game/GameObjects/BrickWall.cpp
@@ -80,6 +80,6 @@ void BrickWall::render() const {
 	renderBrick(EBrickLocation::TopLeft);
 	renderBrick(EBrickLocation::TopRight);
 }
-void BrickWall::update(const uint64_t delta) {
+void BrickWall::update(const double delta) {
 
 }
