+++ src/Game/GameObjects/Water.cpp
@@ -25,6 +25,6 @@ void Water::render() const {
 	renderBlock(EBlockLocation::TopRight);
 }
 
-void Water::update(const uint64_t delta) {
+void Water::update(const double delta) {
 	m_spriteAnimator.update(delta);
 }
