+++ src/Game/Level.cpp
@@ -144,7 +144,7 @@ void Level::render() const {
 		}
 	}
 }
-void Level::update(const uint64_t delta) {
+void Level::update(const double delta) {
 	for (const auto& currentMapObject : m_levelObjects) {
 		if (currentMapObject) {
 			currentMapObject->update(delta);
