+++ src/Resources/ResourceManager.cpp
@@ -241,7 +241,7 @@ bool  ResourceManager::loadJSONRecources(const std::string& JSONPath) {
 				framesDescriptions.reserve(framesArray.Size());
 				for (const auto& currentFrame : framesArray) {
 					const std::string subTextureStr = currentFrame["subTexture"].GetString();
-					const uint64_t duration = currentFrame["duration"].GetUint64();
+					const double duration = currentFrame["duration"].GetDouble();
 					const auto pTextureAtlas = getTexture(textureAtlas);
 					const auto pSubTexture = pTextureAtlas->getSubTexture(subTextureStr);
 					framesDescriptions.emplace_back(pSubTexture.leftBottomUV, pSubTexture.rightTopUV, duration);
